+++ autosubliminal/refiners/namemapping.py
@@ -4,7 +4,7 @@ import logging
 
 from subliminal.video import Episode, Movie
 
-from autosubliminal.util.common import get_alternative_movie_name_mapping, get_alternative_show_name_mapping, \
+from autosubliminal.util.mapping import get_alternative_movie_name_mapping, get_alternative_show_name_mapping, \
     get_movie_name_mapping, get_show_name_mapping
 
 logger = logging.getLogger(__name__)
